Moved classes prefixed Loader under Loader sub-namespace
