+++ src/Loader/ClosureLoader.php
@@ -15,29 +15,36 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Rade\DI\Config\Loader;
+namespace Rade\DI\Loader;
 
+use Rade\DI\AbstractContainer;
+use Rade\DI\Container;
 use Rade\DI\ContainerBuilder;
+use Symfony\Component\Config\Loader\Loader;
 
 /**
- * GlobFileLoader loads files from a glob pattern.
+ * ClosureLoader loads service definitions from a PHP closure.
+ *
+ * The Closure has access to the container as its first argument.
  *
  * @author Divine Niiquaye Ibok <divineibok@gmail.com>
  */
-class GlobFileLoader extends FileLoader
+class ClosureLoader extends Loader
 {
+    /** @var Container|ContainerBuilder */
+    private AbstractContainer $container;
+
+    public function __construct(AbstractContainer $container)
+    {
+        $this->container = $container;
+    }
+
     /**
      * {@inheritdoc}
      */
     public function load($resource, string $type = null): void
     {
-        foreach ($this->glob($resource, false, $globResource) as $path => $info) {
-            $this->import($path);
-        }
-
-        if ($this->container instanceof ContainerBuilder) {
-            $this->container->addResource($globResource);
-        }
+        $resource($this->container);
     }
 
     /**
@@ -45,6 +52,6 @@ class GlobFileLoader extends FileLoader
      */
     public function supports($resource, string $type = null)
     {
-        return 'glob' === $type;
+        return $resource instanceof \Closure;
     }
 }
